refactor fs datasource tests into a mixin
