+++ cpp/src/arrow/dataset/file_parquet_test.cc
@@ -20,18 +20,12 @@
 #include <utility>
 
 #include "arrow/dataset/test_util.h"
-#include "arrow/filesystem/localfs.h"
-#include "arrow/filesystem/path_util.h"
 #include "arrow/record_batch.h"
 #include "arrow/testing/util.h"
-#include "arrow/util/io_util.h"
 #include "parquet/arrow/writer.h"
 
 namespace arrow {
 namespace dataset {
-
-using fs::internal::GetAbstractPathExtension;
-using internal::TemporaryDir;
 
 constexpr int64_t kBatchSize = 1UL << 15;
 constexpr int64_t kBatchRepetitions = 1 << 10;
@@ -94,105 +88,18 @@ TEST_F(TestParquetFileFormat, ScanRecordBatchReader) {
   ASSERT_EQ(row_count, kNumRows);
 }
 
-class TestFileSystemBasedDataSource : public FileSourceFixtureMixin {
- public:
-  void SetUp() override {
-    format_ = std::make_shared<ParquetFileFormat>();
-
-    ASSERT_OK(TemporaryDir::Make("test-fsdatasource-", &temp_dir_));
-    local_fs_ = std::make_shared<fs::LocalFileSystem>();
-
-    auto path = temp_dir_->path().ToString();
-    fs_ = std::make_shared<fs::SubTreeFileSystem>(path, local_fs_);
-
-    CreateEmptyFiles();
+class TestParquetFileSystemBasedDataSource
+    : public FileSystemBasedDataSourceMixin<ParquetFileFormat> {
+  std::vector<std::string> file_names() const override {
+    return {"a/b/c.parquet", "a/b/c/d.parquet", "a/b.parquet", "a.parquet"};
   }
-
-  void CreateFile(std::string path, std::string contents) {
-    auto parent = fs::internal::GetAbstractPathParent(path).first;
-    if (parent != "") {
-      ASSERT_OK(this->fs_->CreateDir(parent, true));
-    }
-    std::shared_ptr<io::OutputStream> file;
-    ASSERT_OK(this->fs_->OpenOutputStream(path, &file));
-    ASSERT_OK(file->Write(contents));
-  }
-
-  void CreateEmptyFiles() {
-    for (auto path : {"a/b/c.parquet", "a/b/c/d.parquet", "a/b.parquet", "a.parquet"}) {
-      CreateFile(path, "");
-    }
-  }
-
-  void MakeDataSource() {
-    ASSERT_OK(FileSystemBasedDataSource::Make(fs_.get(), selector_, format_,
-                                              std::make_shared<ScanOptions>(), &source_));
-  }
-
- protected:
-  fs::Selector selector_;
-  std::unique_ptr<FileSystemBasedDataSource> source_;
-  std::shared_ptr<fs::LocalFileSystem> local_fs_;
-  std::shared_ptr<fs::FileSystem> fs_;
-  std::unique_ptr<TemporaryDir> temp_dir_;
-  std::shared_ptr<ParquetFileFormat> format_;
 };
 
-TEST_F(TestFileSystemBasedDataSource, NonRecursive) {
-  selector_.base_dir = "/";
-  MakeDataSource();
+TEST_F(TestParquetFileSystemBasedDataSource, NonRecursive) { this->NonRecursive(); }
 
-  int count = 0;
-  ASSERT_OK(source_->GetFragments({})->Visit([&](std::shared_ptr<DataFragment> fragment) {
-    auto file_fragment = internal::checked_pointer_cast<FileBasedDataFragment>(fragment);
-    ++count;
-    auto extension =
-        fs::internal::GetAbstractPathExtension(file_fragment->source().path());
-    EXPECT_TRUE(format_->IsKnownExtension(extension));
-    std::shared_ptr<io::RandomAccessFile> f;
-    return this->fs_->OpenInputFile(file_fragment->source().path(), &f);
-  }));
+TEST_F(TestParquetFileSystemBasedDataSource, Recursive) { this->Recursive(); }
 
-  ASSERT_EQ(count, 1);
-}
-
-TEST_F(TestFileSystemBasedDataSource, Recursive) {
-  selector_.base_dir = "/";
-  selector_.recursive = true;
-  MakeDataSource();
-
-  int count = 0;
-  ASSERT_OK(source_->GetFragments({})->Visit([&](std::shared_ptr<DataFragment> fragment) {
-    auto file_fragment = internal::checked_pointer_cast<FileBasedDataFragment>(fragment);
-    ++count;
-    auto extension =
-        fs::internal::GetAbstractPathExtension(file_fragment->source().path());
-    EXPECT_TRUE(format_->IsKnownExtension(extension));
-    std::shared_ptr<io::RandomAccessFile> f;
-    return this->fs_->OpenInputFile(file_fragment->source().path(), &f);
-  }));
-
-  ASSERT_EQ(count, 4);
-}
-
-TEST_F(TestFileSystemBasedDataSource, DeletedFile) {
-  selector_.base_dir = "/";
-  selector_.recursive = true;
-  MakeDataSource();
-  ASSERT_OK(this->fs_->DeleteFile("a/b.parquet"));
-
-  ASSERT_RAISES(
-      IOError,
-      source_->GetFragments({})->Visit([&](std::shared_ptr<DataFragment> fragment) {
-        auto file_fragment =
-            internal::checked_pointer_cast<FileBasedDataFragment>(fragment);
-        auto extension =
-            fs::internal::GetAbstractPathExtension(file_fragment->source().path());
-        EXPECT_TRUE(format_->IsKnownExtension(extension));
-        std::shared_ptr<io::RandomAccessFile> f;
-        return this->fs_->OpenInputFile(file_fragment->source().path(), &f);
-      }));
-}
+TEST_F(TestParquetFileSystemBasedDataSource, DeletedFile) { this->DeletedFile(); }
 
 }  // namespace dataset
 }  // namespace arrow
